refactor(proxy): fold proxy_common into caddy and delete it (#2088)
After #2081 deleted proxy.py, caddy.py was the sole consumer of
proxy_common.py (the transient staging module from #2080). Fold its five
helpers + their deps into caddy.py and delete proxy_common.py — the end
state of the #1642 retirement: no proxy.py and no proxy_common.py;
helpers live in caddy, the sole engine.

Moved into caddy.py: detect_host_ipv4s, _is_loopback (+ _LOOPBACK_NETS),
_FALLBACK_ACL_SUBNETS, _FALLBACK_DENY_SUBNETS, _proxy_preexec (+ the
prctl machinery). test_caddy.py repoints to the caddy-local helpers.

Part 6/5 of the nginx retirement (#1642). Stacks on #2081 (#2091).
